Add a Screen for QR Code Scanning Presentation
