+++ app/src/main/res/layout/alert_message.xml
@@ -1,11 +1,9 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/rlBackground"
-    android:background="@android:color/transparent"
-    tools:context="com.brh.pronapmobile.fragments.AlertFragment">
+    android:background="@android:color/transparent">
 
     <android.support.v7.widget.CardView
         android:layout_height="wrap_content"
